model/requirements.txt
torch
torchvision
numpy
pandas
fastapi
uvicorn
python-multipart
pillow
